Functional d/e system
Full functionality for encryption and decryption given a string. Encrypted messages appear as a series of numbers but will still convert back to the original words when decrypted. I'm assuming this is fine, the point is to encrypt a message, which we have done. The private key (d) also seemed to be the issue with the large ints, using Euc ext. aglo the d value was corrected and now decrypts the code. If the function for the d value is changed back to the pow(p,-1,q) the code will not run.

Finally unless there are getter and setter methods in python that I am unaware of global values will be needed if you want dynamic input from the user (i.e. entering a message) that can be used in other files. Feel free to test the encryption/decryption with different words